refactor(search): Copilot 리뷰 반영 - SliceResponse 패키지 이동 및 입력값 검증 추가
- SliceResponse를 common.web.response에서 common.pagination으로 이동
- /suggestions 엔드포인트 query 파라미터에 @NotBlank 추가
- 검색 목록 엔드포인트 @NotEmpty → @NotBlank 통일
- 전체 pageSize 파라미터에 @Positive 추가
