+++ src/main/java/com/backend/allreva/module/search/application/RentSearchService.java
@@ -1,7 +1,7 @@
 package com.backend.allreva.module.search.application;
 
 import com.backend.allreva.common.exception.CustomException;
-import com.backend.allreva.common.web.response.SliceResponse;
+import com.backend.allreva.common.pagination.SliceResponse;
 import com.backend.allreva.module.search.application.dto.RentThumbnail;
 import com.backend.allreva.module.search.application.port.RentSearchRepository;
 import com.backend.allreva.module.search.exception.SearchErrorCode;
