+++ src/main/java/com/backend/allreva/module/search/application/port/RentSearchRepository.java
@@ -1,6 +1,6 @@
 package com.backend.allreva.module.search.application.port;
 
-import com.backend.allreva.common.web.response.SliceResponse;
+import com.backend.allreva.common.pagination.SliceResponse;
 import com.backend.allreva.module.search.application.dto.RentThumbnail;
 import java.util.List;
