+++ src/main/java/com/backend/allreva/module/search/infra/postgresql/SurveySearchRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.backend.allreva.module.search.infra.postgresql;
 
-import com.backend.allreva.common.web.response.SliceResponse;
+import com.backend.allreva.common.pagination.SliceResponse;
 import com.backend.allreva.module.recruitment.survey.domain.QSurvey;
 import com.backend.allreva.module.recruitment.survey.domain.participant.QSurveyParticipant;
 import com.backend.allreva.module.search.application.dto.SurveyThumbnail;
